+++ src/components/Chips/Chips.jsx
@@ -8,6 +8,7 @@ import { getCSSVar } from "../../services/themes";
 import { NOOP } from "../../utils/function-utils";
 import { elementColorsNames, getElementColor } from "../../utils/colors-vars-map";
 import "./Chips.scss";
+import Avatar from "../Avatar/Avatar";
 
 const Chips = forwardRef(
   (
@@ -17,6 +18,8 @@ const Chips = forwardRef(
       label,
       leftIcon,
       rightIcon,
+      leftAvatar,
+      rightAvatar,
       disabled,
       readOnly,
       allowTextSelection,
@@ -61,6 +64,16 @@ const Chips = forwardRef(
         onMouseDown={onMouseDown}
         data-testid={dataTestId}
       >
+        {leftAvatar ? (
+          <Avatar
+            withoutBorder
+            className="chip-avatar left"
+            customSize={16}
+            src={leftAvatar}
+            type={Avatar.types.IMG}
+            key={id}
+          />
+        ) : null}
         {leftIcon ? (
           <Icon
             className="chip-icon left"
@@ -80,6 +93,16 @@ const Chips = forwardRef(
             icon={rightIcon}
             iconSize={iconSize}
             ignoreFocusStyle
+          />
+        ) : null}
+        {rightAvatar ? (
+          <Avatar
+            withoutBorder
+            className="chip-avatar right"
+            customSize={16}
+            src={rightAvatar}
+            type={Avatar.types.IMG}
+            key={id}
           />
         ) : null}
         {hasCloseButton && (
@@ -111,6 +134,10 @@ Chips.propTypes = {
   rightIcon: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
   /** Icon to place on the left */
   leftIcon: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
+  /** Img to place as avatar on the right */
+  rightAvatar: PropTypes.string,
+  /** Img to place as avatar on the left */
+  leftAvatar: PropTypes.string,
   color: PropTypes.oneOf(Object.keys(Chips.colors)),
   /** size for font icon */
   iconSize: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
@@ -136,6 +163,8 @@ Chips.defaultProps = {
   readOnly: false,
   rightIcon: null,
   leftIcon: null,
+  leftAvatar: null,
+  rightAvatar: null,
   color: Chips.colors.PRIMARY,
   iconSize: 16,
   onDelete: NOOP,