+++ src/components/TextWithHighlight/TextWithHighlight.jsx
@@ -1,6 +1,8 @@
-import React, { useRef, forwardRef, useMemo, useEffect } from "react";
+import React, { useRef, forwardRef, useMemo, useLayoutEffect } from "react";
 import PropTypes from "prop-types";
 import cx from "classnames";
+import Tooltip from "components/Tooltip/Tooltip";
+import useIsOverflowing from "hooks/useIsOverflowing";
 import useMergeRefs from "../../hooks/useMergeRefs";
 import "./TextWithHighlight.scss";
 
@@ -15,42 +17,82 @@ const getTextPart = (text, key, shouldHighlight) => {
   return <span key={key}>{text}</span>;
 };
 
-const TextWithHighlight = forwardRef(({ className, id, text, highlightTerm, limit, linesToClamp }, ref) => {
-  const componentRef = useRef(null);
-  const mergedRef = useMergeRefs({ refs: [ref, componentRef] });
+const TextWithHighlight = forwardRef(
+  (
+    {
+      className,
+      id,
+      text,
+      highlightTerm,
+      limit,
+      useEllipsis,
+      linesToClamp,
+      ignoreCase,
+      allowTermSplit,
+      nonEllipsisTooltip,
+      tooltipPosition
+    },
+    ref
+  ) => {
+    const componentRef = useRef(null);
+    const mergedRef = useMergeRefs({ refs: [ref, componentRef] });
 
-  const textWithHighlights = useMemo(() => {
-    if (!text || !highlightTerm || limit === 0) return text;
-    const tokens = text.split(new RegExp(`(${highlightTerm})`, "i")); // ignore case
-    const parts = [];
-    // Tokens include the term search (in odd indices)
-    let highlightTermsCount = 0;
-    let key = 0;
-    for (let i = 0; i < tokens.length; i++) {
-      // skip empty tokens
-      if (tokens[i]) {
-        // adding highlightend part
-        const isTermPart = i % 2 === 1;
-        const shouldHighlight = isTermPart && (!limit || limit < 0 || highlightTermsCount < limit);
-        parts.push(getTextPart(tokens[i], key++, shouldHighlight));
-        if (isTermPart) highlightTermsCount++;
+    const textWithHighlights = useMemo(() => {
+      if (!text || !highlightTerm || limit === 0) return text;
+      let finalTerm = highlightTerm;
+      if (allowTermSplit) {
+        finalTerm = highlightTerm.split(" ").join("|");
       }
+      const tokens = text.split(new RegExp(`(${finalTerm})`, ignoreCase ? "i" : ""));
+      const parts = [];
+      // Tokens include the term search (in odd indices)
+      let highlightTermsCount = 0;
+      let key = 0;
+      for (let i = 0; i < tokens.length; i++) {
+        // skip empty tokens
+        if (tokens[i]) {
+          // adding highlight part
+          const isTermPart = i % 2 === 1;
+          const shouldHighlight = isTermPart && (!limit || limit < 0 || highlightTermsCount < limit);
+          parts.push(getTextPart(tokens[i], key++, shouldHighlight));
+          if (isTermPart) highlightTermsCount++;
+        }
+      }
+
+      return parts;
+    }, [text, highlightTerm, limit, ignoreCase, allowTermSplit]);
+
+    const isOverflowing = useIsOverflowing({ ref: useEllipsis && componentRef });
+
+    useLayoutEffect(() => {
+      if (componentRef.current) {
+        componentRef.current.style.setProperty("--heading-clamp-lines", linesToClamp);
+      }
+    }, [componentRef, linesToClamp]);
+
+    let Element = (
+      <div
+        ref={mergedRef}
+        className={cx("text-with-highlight--wrapper", className, {
+          "with-ellipsis": useEllipsis
+        })}
+        id={id}
+      >
+        {textWithHighlights}
+      </div>
+    );
+
+    if (isOverflowing || nonEllipsisTooltip) {
+      const tooltipContent = isOverflowing ? text : nonEllipsisTooltip;
+      return (
+        <Tooltip content={tooltipContent} position={tooltipPosition}>
+          {Element}
+        </Tooltip>
+      );
     }
-
-    return parts;
-  }, [text, highlightTerm, limit]);
-
-  useEffect(() => {
-    if (!componentRef.current) return;
-    componentRef.current.style.setProperty("-webkit-line-clamp", linesToClamp);
-  }, [componentRef, linesToClamp]);
-
-  return (
-    <div ref={mergedRef} className={cx("text-with-highlight--wrapper", className)} id={id}>
-      {textWithHighlights}
-    </div>
-  );
-});
+    return Element;
+  }
+);
 
 TextWithHighlight.propTypes = {
   className: PropTypes.string,
@@ -58,7 +100,14 @@ TextWithHighlight.propTypes = {
   text: PropTypes.string,
   highlightTerm: PropTypes.string,
   limit: PropTypes.number,
-  linesToClamp: PropTypes.number
+  ignoreCase: PropTypes.bool,
+  /** Should use ellipsis */
+  useEllipsis: PropTypes.bool,
+  /** Allow highlight every word as a separate term */
+  allowTermSplit: PropTypes.bool,
+  linesToClamp: PropTypes.number,
+  /** Tooltip to show when there is no overflow */
+  nonEllipsisTooltip: PropTypes.string
 };
 
 TextWithHighlight.defaultProps = {
@@ -66,8 +115,12 @@ TextWithHighlight.defaultProps = {
   id: undefined,
   text: "",
   highlightTerm: null,
+  allowTermSplit: true,
   limit: null,
-  linesToClamp: 3
+  ignoreCase: true,
+  useEllipsis: true,
+  linesToClamp: 3,
+  nonEllipsisTooltip: null
 };
 
 export default TextWithHighlight;